Safari-609.1.11.1/LayoutTests/inspector/dom-debugger/event-timeout-breakpoints-expected.txt

Tests for Event Timer breakpoints.


== Running test suite: DOMDebugger.Event.Timer
-- Running test case: DOMDebugger.Event.Timeout.AddBreakpoint
Adding "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleWindow_setTimeout
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] trigger_setTimeout
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.AddDisabledBreakpoint
Adding "timeout" Event Breakpoint...
Disabling "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.RemoveBreakpoint
Adding "timeout" Event Breakpoint...
Removing "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Timeout.RemoveDisabledBreakpoint
Adding "timeout" Event Breakpoint...
Disabling "timeout" Event Breakpoint...
Removing "timeout" Event Breakpoint...
Firing "setTimeout" on window...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.
